+++ internal/provider/resource_computing.go
@@ -319,6 +319,18 @@ func resourceComputingDelete(ctx context.Context, d *schema.ResourceData, meta i
 	if resourceId == "" {
 		return diag.Errorf(COMPUTING_ID + " not found! ")
 	}
+
+	resp1, r1, err1 := apiClient.CoreServiceApi.StopService(ctx, resourceId).Execute()
+	if err1 != nil {
+		return diag.Errorf("Error when calling computing `CoreServiceApi.StopService``: %v\n", err1)
+	}
+	if r1.StatusCode != 200 {
+		return diag.Errorf("Stop resource computing fail with %d . ", r1.StatusCode)
+	}
+	if errRefresh := waitJobComplete(ctx, apiClient.CoreJobServiceApi, resp1.GetId()); errRefresh != nil {
+		return diag.Errorf(errRefresh.Error())
+	}
+
 	resp, r, err := apiClient.CoreServiceApi.DeleteService(ctx, resourceId).Execute()
 	if err != nil {
 		return diag.Errorf("Error when calling `CoreServiceApi.DeleteService``: %v\n", err)
